Use new navbar everywhere.
